Alterado o tipo de dado de double pra int em onProgressUpdated

+++ app/src/main/java/br/com/bg7/appvistoria/data/source/remote/HttpProgressCallback.java
@@ -5,5 +5,5 @@ package br.com.bg7.appvistoria.data.source.remote;
  * Date: 2017-07-28
  */
 public interface HttpProgressCallback<T> extends HttpCallback<T>{
-    void onProgressUpdated(int percentage);
+    void onProgressUpdated(double percentage);
 }
